zmiana metod dla postow, mapowanie na dto

+++ BUSINESS/src/main/java/com/example/facebookapi/service/PostService.java
@@ -1,11 +1,14 @@
 package com.example.facebookapi.service;
 
+import com.example.facebookapi.dto.PostDto;
 import com.example.facebookapi.entity.Comment;
 import com.example.facebookapi.entity.Post;
 import com.example.facebookapi.entity.User;
 import com.example.facebookapi.exceptions.PostException;
 import com.example.facebookapi.exceptions.PostNotExist;
 import com.example.facebookapi.exceptions.UserNotExist;
+import com.example.facebookapi.exceptions.UsernameNotExist;
+import com.example.facebookapi.mappers.PostMapper;
 import com.example.facebookapi.repository.CommentRepository;
 import com.example.facebookapi.repository.PostRepository;
 import com.example.facebookapi.repository.UserRepository;
@@ -22,33 +25,42 @@ public class PostService {
     private final CommentService commentService;
     private final CommentRepository commentRepository;
     private final UserRepository userRepository;
+    private final PostMapper postMapper;
 
 
-        public Post savePost (Post post) {
+        public PostDto savePost (PostDto postDto) {
+            Optional<User> userFromDB = userRepository.findByUserName(postDto.getUserName());
+            if (userFromDB.isEmpty()) {
+                throw new UsernameNotExist(postDto.getUserName());
+            }
 
-
-            if ((post.getDescription() == null || post.getDescription().isBlank()) &&
-                (post.getPostImgURL() == null || post.getPostImgURL().isBlank())) {
+            if ((postDto.getDescription() == null || postDto.getDescription().isBlank()) &&
+                (postDto.getPostImgURL() == null || postDto.getPostImgURL().isBlank())) {
             throw new PostException();
         }
 
 
         LocalDateTime time = LocalDateTime.now();
 
-        post.setPostID(UUID.randomUUID());
-        post.setLikes(0);
-        post.setDateTime(time);
+            Post newPost = postMapper.dtoToPost(postDto);
+            newPost.setPostID(UUID.randomUUID());
+            newPost.setUserID(userFromDB.get().getUserID());
+            newPost.setImageURL(userFromDB.get().getUserImage());
+            newPost.setLikes(0);
+            newPost.setDateTime(time);
+            postRepository.save(newPost);
 
-        return postRepository.save(post);
+            return postMapper.toPostDto(newPost);
 
     }
 
 
-    public List<Post> getPosts(){
-        return postRepository.findAll();
+    public List<PostDto> getPosts(){
+        List<Post> posts = postRepository.findAll();
+        return postMapper.toPostDtos(posts);
     }
 
-    public List<Post> deletePost(UUID postID){
+    public List<PostDto> deletePost(UUID postID){
         Optional<Post> post = postRepository.findById(postID);
         if (post.isEmpty()){
             throw new PostNotExist(postID);
@@ -61,25 +73,30 @@ public class PostService {
         return getPosts();
     }
 
-    public List<Post> getUserPosts (UUID userID){
+    public List<PostDto> getUserPosts (UUID userID){
         Optional<User> userFromDB = userRepository.findById(userID);
         if (userFromDB.isEmpty()) {
             throw new UserNotExist(userID);
         }
 
-        return postRepository.findAllByUserID(userID);
+        List<Post> userPosts = postRepository.findAllByUserID(userID);
+        return postMapper.toPostDtos(userPosts);
     }
 
-    public List<Post> deleteUserPosts (UUID userID){
-      //  checkUser(userID);
-        List<Post> toDelete = postRepository.findAllByUserID(userID);
+    public List<PostDto> deleteUserPosts (UUID userID){
+        Optional<User> userFromDB = userRepository.findById(userID);
+        if (userFromDB.isEmpty()) {
+            throw new UserNotExist(userID);
+        }
 
-        toDelete.forEach(post -> {
+        List<Post> userPosts = postRepository.findAllByUserID(userID);
+
+        userPosts.forEach(post -> {
             List<Comment> commentsToDelete = commentService.getCommentsByPostID(post.getPostID());
-            commentRepository.deleteAll(commentsToDelete);
-        });
 
-        postRepository.deleteAll(toDelete);
+            commentRepository.deleteAll(commentsToDelete);
+            postRepository.delete(post);
+        });
 
         return getPosts();
     }
